dgen/requirements.txt
nltk==3.4.3
pyspellchecker==0.5.0 